Fix of some issues
